requirements/base.txt
Django==1.11.6
django-bootstrap3==11.0.0
sqlparse==0.1.19
django-registration-redux==1.4
httplib2==0.9.1
oauth2==1.5.211
oauthlib==1.0.3
Pillow==3.3.2
python-jose==3.0.1
python-openid==2.2.5
python-social-auth==0.3.6
social-auth-app-django
requests==2.9.1
requests-oauthlib==0.6.1
Unidecode==0.4.18
django-bower==5.2.0
Paver==1.2.4
psutil==1.2.1
lazy==1.1
path.py==8.2.1
watchdog==0.7.1
python-memcached
libsass==0.10.0
paved==0.5.1
django-debug-toolbar==1.4
djangorestframework
pytz
